[뒤끝서버] UserData { 재화, 레벨, 경험치 } 작업

+++ P1/Assets/@Scripts/UI/SubItem/Item/UI_GoodItem.cs
@@ -48,7 +48,7 @@ public class UI_GoodItem : UI_Base
         if (goodType == EGoodType.None)
             return;
 
-        GetTMPText((int)Texts.GoodText).text = Util.ConvertToKoreanUnits(Managers.Purse.GetAmount(goodType));
+        GetTMPText((int)Texts.GoodText).text = Util.ConvertToKoreanUnits((long)Managers.Purse.GetAmount(goodType));
     }
 
     public RectTransform GetGoodIcon()
